Fix layout of checkbox elements for Bootstrap 4
Fixes #715
