--- addon/components/base/bs-form/element/layout/horizontal/checkbox.js
@@ -1,13 +1,0 @@
-import FormElementLayoutHorizontal from '../horizontal';
-import layout from 'ember-bootstrap/templates/components/bs-form/element/layout/horizontal/checkbox';
-
-/**
-
- @class FormElementLayoutHorizontalCheckbox
- @namespace Components
- @extends Components.FormElementLayout
- @private
- */
-export default FormElementLayoutHorizontal.extend({
-  layout
-});
