+++ addon/components/bs4/bs-form/element/layout/horizontal/checkbox.js
@@ -1,1 +1,13 @@
-export { default } from 'ember-bootstrap/components/base/bs-form/element/layout/horizontal/checkbox';
+import FormElementLayoutHorizontal from '../horizontal';
+import layout from 'ember-bootstrap/templates/components/bs-form/element/layout/horizontal/checkbox';
+
+/**
+
+ @class FormElementLayoutHorizontalCheckbox
+ @namespace Components
+ @extends Components.FormElementLayout
+ @private
+ */
+export default FormElementLayoutHorizontal.extend({
+  layout
+});
