+++ addon/templates/components/bs4/bs-form/element/layout/horizontal/checkbox.hbs
@@ -1,6 +1,8 @@
 <div class="{{horizontalInputGridClass}} {{horizontalInputOffsetGridClass}}">
-  {{yield}}
-  {{component labelComponent}}
-  {{component errorsComponent}}
-  {{component helpTextComponent}}
+  <div class="form-check">
+    {{yield}}
+    {{component labelComponent}}
+    {{component errorsComponent}}
+    {{component helpTextComponent}}
+  </div>
 </div>
