+++ addon/templates/components/bs4/bs-form/element/layout/vertical/checkbox.hbs
@@ -1,4 +1,6 @@
-{{yield}}
-{{component labelComponent}}
-{{component errorsComponent}}
-{{component helpTextComponent}}
+<div class="form-check">
+  {{yield}}
+  {{component labelComponent}}
+  {{component errorsComponent}}
+  {{component helpTextComponent}}
+</div>
